refactor: remove quantum extension
move test operations to test-only extension
