+++ src/hugr/rewrite/simple_replace.rs
@@ -234,9 +234,9 @@ pub(in crate::hugr::rewrite) mod test {
     use crate::ops::OpTag;
     use crate::ops::{OpTrait, OpType};
     use crate::std_extensions::logic::test::and_op;
-    use crate::std_extensions::quantum::test::{cx_gate, h_gate};
     use crate::type_row;
     use crate::types::{FunctionType, Type};
+    use crate::utils::test_quantum_extension::{cx_gate, h_gate};
     use crate::{IncomingPort, Node};
 
     use super::SimpleReplacement;
