+++ src/std_extensions.rs
@@ -5,4 +5,3 @@
 pub mod arithmetic;
 pub mod collections;
 pub mod logic;
-pub mod quantum;
